use lsregister to register the shortcut.
This isn't necessary but processing a large number of shortcuts in a directory that isn't immediately detected will not register the app with launch services. This could cause excessive processing by nesessionmanager (network extension). It detects the app before it's registered and loops endlessly trying to find the shortcut.app in launch services database.